requirements.txt
fastapi
uvicorn
transformers
torch
accelerate
pydantic
scipy
python-dotenv
requests
